+++ CHANGELOG.md
@@ -8,6 +8,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Ability to remove topics from a cluster.
 - Ability to search topics within a cluster.
 
+## 2.1.4 (02/19/2019)
+- [ISSUE-136](https://github.com/SourceLabOrg/kafka-webview/issues/136) Fix URLs when running Kafka-Webview behind a reverse proxy with a URL prefix.  You can configure Kafka WebView by setting the following configuration option in your config.yml file:
+- Updated Kafka Client library version from 2.0.0 to 2.0.1
+
+```yml
+server:
+  servlet:
+    context-path: /prefix/here
+```
+>>>>>>> 09b13e34f4aa14c65afb9d5fcd207b560e8e1a23
+
 ## 2.1.3 (01/19/2019)
 
 #### Bug fixes